Factor out app restore selection code into new AppSelectionManager
